dcm-exam/requirements.txt
pywebview==5.1
pydicom==2.4.4
